requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-dotenv==1.0.1
httpx==0.28.1
openai==1.82.0
google-generativeai==0.8.4
yfinance==0.2.51
apscheduler==3.10.4
pydantic==2.10.4
lxml==5.3.0
aiosqlite==0.20.0
